# swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Please read on https://nordkern.dk/wp-content/uploads/2018/03/Projektplan.pdf",
    "version" : "0.0.1",
    "title" : "soeofficer",
    "termsOfService" : "Please read on https://nordkern.dk/wp-content/uploads/2018/03/Projektplan.pdf",
    "contact" : {
      "name" : "Nordkern ApS",
      "url" : "https://nordkern.dk",
      "email" : "kontakt@nordkern.dk"
    }
  },
  "basePath" : "/api",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/officer/search" : {
      "post" : {
        "tags" : [ "officer" ],
        "summary" : "Get all officers matching the supplied criteria",
        "description" : "",
        "operationId" : "getPersonsMatchingCriteria",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "givenName",
          "in" : "query",
          "description" : "The given name of the person",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "surname",
          "in" : "query",
          "description" : "The surname of the person",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "dateOfBirth",
          "in" : "query",
          "description" : "The person's date of birth. Format is: DD/MM/YYYY",
          "required" : false,
          "type" : "date"
        }, {
          "name" : "gender",
          "in" : "query",
          "description" : "The gender of the person. Accepted values are from the set {Mand,Kvinde,Ukendt}",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "appointedDate",
          "in" : "query",
          "description" : "The date from which the officer title is valid",
          "required" : false,
          "type" : "date"
        }, {
          "name" : "rankID",
          "in" : "query",
          "description" : "The ID of the rank to be used as a criteria",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "dodabNumber",
          "in" : "query",
          "description" : "The officer's Dodab number",
          "required" : false,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Officer"
              }
            }
          }
        }
      }
    },
...
    "/person" : {
      "post" : {
        "tags" : [ "person" ],
        "summary" : "Add a new person to the system",
        "description" : "",
        "operationId" : "createPerson",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "givenName",
          "in" : "query",
          "description" : "The given name of the person",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "surname",
          "in" : "query",
          "description" : "The surname of the person",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "dateOfBirth",
          "in" : "query",
          "description" : "The person's date of birth. Format is: DD/MM/YYYY",
          "required" : true,
          "type" : "date"
        }, {
          "name" : "gender",
          "in" : "query",
          "description" : "The gender of the person. Accepted values are from the set {Mand,Kvinde,Ukendt}",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "dateOfDeath",
          "in" : "query",
          "description" : "The date the person has dies",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Person"
            }
          }
        }
      }
    },
...
    "/person/search" : {
      "post" : {
        "tags" : [ "person" ],
        "summary" : "Get all persons in the system matching the supplied criteria",
        "description" : "",
        "operationId" : "getPersonsMatchingCriteria",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "givenName",
          "in" : "query",
          "description" : "The given name of the person",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "surname",
          "in" : "query",
          "description" : "The surname of the person",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "dateOfBirth",
          "in" : "query",
          "description" : "The person's date of birth. Format is: DD/MM/YYYY",
          "required" : false,
          "type" : "date"
        }, {
          "name" : "gender",
          "in" : "query",
          "description" : "The gender of the person. Accepted values are from the set {Mand,Kvinde,Ukendt}",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Person"
              }
            }
          }
        }
      }
    },
...
    "/person/{id}" : {
      "get" : {
        "tags" : [ "person" ],
        "summary" : "Gets specific person by id",
        "description" : "",
        "operationId" : "readPerson",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Person"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "person" ],
        "summary" : "Update an existing person of the system",
        "description" : "",
        "operationId" : "updatePerson",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "givenName",
          "in" : "query",
          "description" : "The given name of the person",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "surname",
          "in" : "query",
          "description" : "The surname of the person",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "dateOfBirth",
          "in" : "query",
          "description" : "The person's date of birth. Format is: DD/MM/YYYY",
          "required" : false,
          "type" : "date"
        }, {
          "name" : "gender",
          "in" : "query",
          "description" : "The gender of the person. Accepted values are from the set {Mand,Kvinde,Ukendt}",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "dateOfDeath",
          "in" : "query",
          "description" : "The date the person has dies",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Person"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "person" ],
        "summary" : "Delete person from the system",
        "description" : "",
        "operationId" : "deletePerson",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/person/{id}/relations" : {
      "get" : {
        "tags" : [ "person" ],
        "summary" : "Gets parents of person",
        "description" : "",
        "operationId" : "getRelations",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/RelationsResponse"
            }
          }
        }
      }
    },
...
    "/user/search" : {
      "post" : {
        "tags" : [ "user" ],
        "summary" : "Get all users in the system matching the supplied criteria",
        "description" : "",
        "operationId" : "getPersonsMatchingCriteria",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "username",
          "in" : "query",
          "description" : "The username of the user",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "email",
          "in" : "query",
          "description" : "The email of the user",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/User"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
    "Officer" : {
      "type" : "object",
      "required" : [ "appointedNumber", "dodabNumber", "id", "person", "terminationCause" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 1,
          "description" : "The unique ID of the officer"
        },
        "appointedNumber" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 12,
          "description" : "The appointed number of the officer"
        },
        "dodabNumber" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 123,
          "description" : "The officer's Dodab number"
        },
        "appointedDate" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The date from which the officer title is valid",
          "readOnly" : true
        },
        "appointedUntil" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The date from which the officer title is invalid"
        },
        "terminationCause" : {
          "type" : "string",
          "example" : "Accidental_death",
          "description" : "The cause of termination",
          "enum" : [ "Terminated", "Killed_in_action", "Accidental_death", "Transferred", "Other" ]
        },
        "person" : {
          "description" : "The person associated with the officer",
          "$ref" : "#/definitions/Person"
        }
      }
    },
...
    "Person" : {
      "type" : "object",
      "required" : [ "dateOfBirth", "gender", "givenName", "id", "surname" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 1,
          "description" : "The unique ID of the person"
        },
        "givenName" : {
          "type" : "string",
          "example" : "John",
          "description" : "The given name of the person"
        },
        "surname" : {
          "type" : "string",
          "example" : "Doe",
          "description" : "The surname of the person"
        },
        "dateOfBirth" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The person's date of birth. Format is: DD/MM/YYYY"
        },
        "gender" : {
          "type" : "string",
          "example" : "Mand",
          "description" : "The gender of the person. Accepted values are from the set {Mand,Kvinde,Ukendt}",
          "enum" : [ "Male", "Female", "Unknown" ]
        },
        "dateOfDeath" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The date of death. Format is: DD/MM/YYYY"
        }
      }
    },
    "PersonRelation" : {
      "type" : "object",
      "required" : [ "dateOfBirth", "gender", "givenName", "id", "surname" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 1,
          "description" : "The unique ID of the person"
        },
        "givenName" : {
          "type" : "string",
          "example" : "John",
          "description" : "The given name of the person"
        },
        "surname" : {
          "type" : "string",
          "example" : "Doe",
          "description" : "The surname of the person"
        },
        "dateOfBirth" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The person's date of birth. Format is: DD/MM/YYYY"
        },
        "gender" : {
          "type" : "string",
          "example" : "Mand",
          "description" : "The gender of the person. Accepted values are from the set {Mand,Kvinde,Ukendt}",
          "enum" : [ "Male", "Female", "Unknown" ]
        },
        "dateOfDeath" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The date of death. Format is: DD/MM/YYYY"
        },
        "type" : {
          "type" : "string",
          "example" : "{mother,father}",
          "description" : "The type of parent",
          "enum" : [ "Mother", "Father" ]
        },
        "isOfficer" : {
          "type" : "boolean",
          "example" : false,
          "description" : "If the relation is an officer"
        },
        "relationID" : {
          "type" : "integer",
          "format" : "int64",
          "example" : "true / false",
          "description" : "The id of the relation"
        }
      }
    },
...
    "RelationsResponse" : {
      "type" : "object",
      "required" : [ "children", "parents" ],
      "properties" : {
        "parents" : {
          "type" : "array",
          "example" : "list",
          "description" : "The parents of the relation",
          "items" : {
            "$ref" : "#/definitions/PersonRelation"
          }
        },
        "children" : {
          "type" : "array",
          "example" : "list",
          "description" : "The children of the relation",
          "items" : {
            "$ref" : "#/definitions/PersonRelation"
          }
        }
      }
    },
    "User" : {
      "type" : "object",
      "required" : [ "email", "from", "id", "role", "until", "username" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 1,
          "description" : "The unique ID of the user"
        },
        "username" : {
          "type" : "string",
          "example" : "john_doe",
          "description" : "The unique username of the user"
        },
        "email" : {
          "type" : "string",
          "example" : "john@doe.com",
          "description" : "The unique email of the user"
        },
        "from" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The date from which the user is valid. Format is: DD/MM/YYYY"
        },
        "until" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "31/12/9999",
          "description" : "The date on which to user is invalid. Format is: DD/MM/YYYY"
        },
        "role" : {
          "type" : "string",
          "example" : "administrator",
          "description" : "The system-role of the user",
          "enum" : [ "contributor", "administrator", "read", "privileged_read" ]
        }
      }
    }
  }
}